Task6

+++ 0x00-hello_world/6-size.c
@@ -14,10 +14,10 @@ int i;
 long int j;
 long long int k;
 float f;
-printf("Size of a char: %ld byte(s)\n", (unsigned long)sizeof(c));
-printf("Size of an int: %ld byte(s)\n", (unsigned long)sizeof(i));
-printf("Size of a long int: %ld byte(s)\n", (unsigned long)sizeof(j));
-printf("Size of a long long int: %ld byte(s)\n", (unsigned long)sizeof(k));
-printf("Size of a float: %ld byte(s)\n", (unsigned long)sizeof(f));
+printf("Size of a char: %ld byte(s)\n",sizeof(c));
+printf("Size of an int: %ld byte(s)\n",sizeof(i));
+printf("Size of a long int: %ld byte(s)\n",sizeof(j));
+printf("Size of a long long int: %ld byte(s)\n",sizeof(k));
+printf("Size of a float: %ld byte(s)\n",sizeof(f));
 return (0);
 }
